Reverted code back to 'global' scope; else was not actually working

+++ .config/bash/ssh_agent.sh
@@ -3,38 +3,20 @@
 # Forward SSH auth requests to Windows or fallback to Keychain
 # If neither are available, warn user
 
-relay() {
-    local listen relay
-
-    listen="UNIX-LISTEN:${SSH_AUTH_SOCK},fork"
-    relay="EXEC:'npiperelay.exe -ei -s //./pipe/openssh-ssh-agent',nofork"
-
-    (command -v "socat" && command -v "npiperelay.exe") &>/dev/null || return
-    pgrep --full 'npiperelay\.exe.*openssh-ssh-agent' &>/dev/null && return
-    rm --force "${SSH_AUTH_SOCK}"
-    (setsid socat "${listen}" "${relay}" & disown) &>/dev/null
-}
-
-kc() {
-    local kc
-
-    command -v "keychain" &>/dev/null || return
-
-    kc="$(keychain \
-        --eval \
-        --agents "ssh" \
+if (command -v socat && command -v npiperelay.exe) &>/dev/null; then
+    if ! pgrep --full 'npiperelay\.exe.*openssh-ssh-agent' &>/dev/null; then
+        [[ -S "${SSH_AUTH_SOCK}" ]] && rm --force "${SSH_AUTH_SOCK}"
+        listen="UNIX-LISTEN:${SSH_AUTH_SOCK},fork"
+        relay="EXEC:'npiperelay.exe -ei -s //./pipe/openssh-ssh-agent',nofork"
+        (setsid socat "${listen}" "${relay}" & disown) &>/dev/null
+        unset listen relay
+    fi
+elif command -v keychain &>/dev/null; then
+    eval "$(keychain --eval --agents 'ssh' \
         "${SSH_HOME}/id_rsa" \
         "${SSH_HOME}/github/id_rsa" \
         "${SSH_HOME}/gitlab/id_rsa" \
     )"
-
-    eval "${kc}"
-}
-
-warn() {
-    printf '%s\n' \
-        "$(tput setaf 3)Failed to configure ssh-agent$(tput sgr0)" >&2
-}
-
-relay || kc || warn
-unset -f relay kc warn
+else
+    printf '%s\n' "Failed to configure ssh-agent" >&2
+fi
